Review: Rename lib/apple to lib/darwin
