+++ lib/darwin/pub_key.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package apple
+package darwin
 
 import (
 	"crypto/ecdsa"
